+++ examples/from-readme-standalone.test.js
@@ -4,7 +4,7 @@ const diskStore = new DiskStore({
     path: __dirname + '/diskcache-standalone', // path for cached files (default: cache)
     ttl: 60 * 60 * 1000, // time to life in milliseconds
                          // (default: never expires)
-    zip: true, // zip files to save disk space (default: false)
+    zip: false, // zip files to save disk space (default: false)
 });
 
 (async () => {